requirements.txt
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
pandas==2.2.1
psycopg2==2.9.9
pydantic==2.5.1
python-telegram-bot==20.6
python-dotenv==1.0.0
SQLAlchemy==2.0.23
matplotlib==3.8.4